requirements.txt
amqp==5.0.9
appdirs==1.4.4
asgiref==3.5.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==22.1.0
bs4==0.0.1
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-filter==2.3.0
django-timezone-field==4.2.3
djangorestframework==3.11.0
fake-useragent==0.1.11
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
kombu==5.2.3
lxml==4.7.1
mypy-extensions==0.4.3
packaging==21.3
parse==1.19.0
pathspec==0.9.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
pyee==8.2.2
pyparsing==3.0.7
pyppeteer==1.0.2
pyquery==1.4.3
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.2
requests==2.27.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.0
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
zipp==3.7.0
